Migrated to QT & qtwebengine, from webkitgtk
This was initially done due to graphical artifacting with webkitgtk,
which when testing with gnome's browser stuff presented the same
behaviour which to me indicated the issue was upstream. The issue I
found was that videos could not perfectly loop, it'd go black just
before the loop breaking it. Hence, we've migrated to qtwebengine and as
a byproduct my gpu usage has dropped significantly, making it even
lower. RAM usage is slightly higher, but that's what you can expect for
chromium. There are also less deps and the build time is smaller, so all
in all a good refactor.